% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\name{adv2ncdf}
\alias{adv2ncdf}
\title{Save an adv object to a NetCDF file}
\usage{
adv2ncdf(x, varTable = NULL, ncfile = NULL, force_v4 = TRUE, debug = 0)
}
\arguments{
\item{x}{an oce object of class \code{adv}, as created by e.g. \code{\link[oce:read.adv]{oce::read.adv()}}.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{force_v4}{logical value which controls the netCDF file version during
the \link[ncdf4]{nc_create} step. The default here is TRUE, whereas the
\link[ncdf4]{ncdf4-package} defaults to FALSE (ensuring that the netCDF
file is compatible with netCDF v3). Some features, including large data
sizes, may require v4.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Given an \code{adv} object created by the \code{oce} package, this function
creates a NetCDF file that can later by read by \code{\link[=ncdf2adv]{ncdf2adv()}} to approximately
reproduce the original contents.
}
\details{
Note that \code{\link[=adv2ncdf]{adv2ncdf()}} defaults \code{varTable} to \code{"adv"}.

The entire contents of the metadata slot are saved in the global attribute named
\code{"metadata"}, in a JSON format.  The JSON material is developed with
\code{\link[=metadata2json]{metadata2json()}}, which yields a value that can be decoded with
\code{\link[=json2metadata]{json2metadata()}}.
}
\examples{
library(ocencdf)

# Example with an adv file from oce package
data(adv, package="oce")
summary(adv)
plot(adv)
# Transfer to NetCDF and back to see if results make sense
oce2ncdf(adv, ncfile="adv.nc")
ADV <- ncdf2adv("adv.nc")
summary(ADV)
plot(ADV)

# Remove temporary file
file.remove("adv.nc")

}
\seealso{
Other things related to adv data: 
\code{\link{ncdf2adv}()}
}
\author{
Dan Kelley and Clark Richards
}
\concept{things related to adv data}
