% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd2ncdf}
\alias{ctd2ncdf}
\title{Save a ctd object to a netcdf file}
\usage{
ctd2ncdf(x, varTable = NULL, ncfile = NULL, debug = 0)
}
\arguments{
\item{x}{an oce object of class \code{ctd}, as created by e.g. \code{\link[oce:as.ctd]{oce::as.ctd()}}
or \code{\link[oce:read.ctd]{oce::read.ctd()}}.}

\item{varTable}{character value indicating the variable-naming
scheme to be used. This is provided to \code{\link[=read.varTable]{read.varTable()}}, the
documentation of which explains the possibilities for pre-defined
schemes named \code{"argo"} and \code{"whp"} and user-supplied files of
similar format.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Save a ctd object to a netcdf file
}
\examples{
library(ocencdf)

# example 1: a ctd file with no per-variable QC flags
data(ctd) # from 'oce' package
oce2ncdf(ctd, ncfile="ctd.nc")
CTD <- as.ctd(ncdf2oce("ctd.nc"))
summary(CTD)

# example 2: a ctd file with per-variable QC flags
data(section) # from 'oce' package
stn <- section[["station", 100]] # 100-th station in section, not station '100'
oce2ncdf(stn, ncfile="stn.nc")
STN <- as.ctd(ncdf2oce("stn.nc"))
summary(STN)

# clean up temporary files (to prevent CRAN test failure)
unlink("ctd.nc")
unlink("stn.nc")

}
\seealso{
Other things related to CTD data: 
\code{\link{ncdf2ctd}()}
}
\author{
Dan Kelley
}
\concept{things related to CTD data}
