% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartable.R
\name{getVariableInfo}
\alias{getVariableInfo}
\title{Get the long name of a variable, using varTable}
\usage{
getVariableInfo(name = NULL, varTable = NULL, oce = NULL)
}
\arguments{
\item{name}{character value naming the variable.  Note that numeric suffices
are trimmed, so that e.g. \code{temperature} and \code{temperature2} yield the
same results.  This is because oce handles e.g. two temperature data streams
by naming the second \code{temperature2}.}

\item{varTable}{either a variable table as read by \code{\link[=read.varTable]{read.varTable()}},
or a character string that is to be passed to that function to
create a variable table.}

\item{oce}{(optional) an oce object.  If provided, then an attempt
is made to infer the unit from it.  Otherwise, the returned \code{unit}
entry is an empty string.}
}
\value{
\code{\link[=getVariableInfo]{getVariableInfo()}} returns a list containing \code{name} (the
name as used in argo netcdf files), \code{long_name} (again, as used in
Argo netcdf files, although the usefulness of this is debatable),
\code{standard_name} (not used by \code{\link[=ctd2ncdf]{ctd2ncdf()}} as of now), \code{FillValue}
(used by \code{\link[=ctd2ncdf]{ctd2ncdf()}} for missing values) and, if \code{oce} is provided
and it can be determined, \code{unit} (a character string specifying
the unit).
}
\description{
This is used by e.g. \code{\link[=ctd2ncdf]{ctd2ncdf()}} to determine how to describe the variable in a
particular flavour of netcdf file, as specified by \code{\link[=read.varTable]{read.varTable()}}.
}
\examples{
library(ocencdf)

# Example
data(ctd)
vt <- read.varTable("argo")
getVariableInfo("temperature", vt, ctd)

}
\author{
Dan Kelley
}
