% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartable.R
\name{getVarInfo}
\alias{getVarInfo}
\title{Get information on a variable, using varTable}
\usage{
getVarInfo(name = NULL, varTable = NULL, oce = NULL, debug = 0)
}
\arguments{
\item{name}{character value naming the variable.  If \code{name} is
not a chemical species (e.g. \code{"NO2"}) then trailing numbers are
removed, as a way to handle both e.g. \code{"temperature"} and \code{"temperature2"},
which may both be present for devices with multiple temperature sensors.
The chemical names are recognized by the presence of one of the following
strings: \code{"CO"}, \code{"NO"} or \code{"PO"}.}

\item{varTable}{either a variable table as read by \code{\link[=read.varTable]{read.varTable()}},
or a character string that is to be passed to that function to
create a variable table.}

\item{oce}{(optional) an oce object.  If provided, then an attempt
is made to infer the unit from it.  Otherwise, the returned \code{unit}
entry is an empty string.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=getVarInfo]{getVarInfo()}} returns a list containing \code{name} (the
name as used in argo netcdf files), \code{long_name} (again, as used in
Argo netcdf files, although the usefulness of this is debatable),
\code{standard_name} (not used by \code{\link[=ctd2ncdf]{ctd2ncdf()}} as of now), \code{FillValue}
(used by \code{\link[=ctd2ncdf]{ctd2ncdf()}} for missing values) and, if \code{oce} is provided
and it can be determined, \code{unit} (a character string specifying
the unit).
}
\description{
This is used by e.g. \code{\link[=ctd2ncdf]{ctd2ncdf()}} to determine how to describe the variable in a
particular flavour of netcdf file, as specified by \code{\link[=read.varTable]{read.varTable()}}.
}
\examples{
library(ocencdf)

# Example
data(ctd)
vt <- read.varTable("argo")
getVarInfo("temperature", vt, ctd)

}
\author{
Dan Kelley
}
