% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json2metadata}
\alias{json2metadata}
\title{Convert a JSON string to an oce metadata slot}
\usage{
json2metadata(j)
}
\arguments{
\item{j}{character value, typically the output from \code{\link[=metadata2json]{metadata2json()}}.}
}
\value{
\link{json2metadata} returns a list in the format of a \code{metadata}
slot from an \code{oce} object.
}
\description{
This is the reverse of \code{\link[=metadata2json]{metadata2json()}}, and is used by
\code{\link[=ncdf2oce]{ncdf2oce()}} and related functions.  See \code{\link[=metadata2json]{metadata2json()}}
for some conversions that are done by that function and
then reversed here.
}
\seealso{
Other things relating to serialization: 
\code{\link{metadata2json}()}
}
\author{
Dan Kelley
}
\concept{things relating to serialization}
