% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{metadata2json}
\alias{metadata2json}
\title{Convert an oce metadata slot to JSON}
\usage{
metadata2json(m, digits = 15)
}
\arguments{
\item{m}{contents of the \code{metadata} slot of an oce object.}

\item{digits}{integer, the number of digits to store in the JSON
representation.}
}
\value{
\link{metadata2json} returns a character value holding
the \code{metadata} slot in JSON, transformed as indicated in
the \dQuote{Details} section.
}
\description{
This converts the contents of an object's \code{metadata} slot
into JSON format, using the \code{toJSON()} function from the
\code{jsonlite} library.  Some conversion is required, and this
is also handled if the results are reconstituted using
\code{\link[=json2metadata]{json2metadata()}}.  See \dQuote{Details}.
}
\details{
Some \code{metadata} items cannot be handled by \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}, so
these are transformed as follows before storage in the return value
NOTE: the reverse function, \code{\link[=json2metadata]{json2metadata()}}, reverses all of these
transformations, so that the original \code{metadata} can be recovered.
\enumerate{
\item Oce uses \code{expression} objects to store units, and these are converted
to character values using \code{\link[=as.character]{as.character()}} before converting to JSON.
\item Oce stores a raw matrix called \code{codes} for \code{adp} objects created from
RDI/Teledyne files, and so this is converted to an integer matrix
before converting to JSON.
\item The following items are converted from POSIXct values to character
values: \code{date}, \code{endTime}, \code{startTime}, and \code{systemUploadTime}.
}
}
\examples{
# Example 1: CTD data
data("ctd")
metadata2json(ctd@metadata)

# Example 2: ADP data
data("adp")
metadata2json(adp@metadata)

}
\seealso{
Other things relating to serialization: 
\code{\link{json2metadata}()}
}
\author{
Dan Kelley
}
\concept{things relating to serialization}
