% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{ncdf2adp}
\alias{ncdf2adp}
\title{Read a netcdf file and create an adp object}
\usage{
ncdf2adp(ncfile = NULL, varTable = NULL, debug = 0)
}
\arguments{
\item{ncfile}{character value naming the input file.}

\item{varTable}{character value indicating the variable-naming
scheme to be used. This is provided to \code{\link[=read.varTable]{read.varTable()}}, the
documentation of which explains the possibilities for pre-defined
schemes named \code{"argo"} and \code{"whp"} and user-supplied files of
similar format.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=ncdf2adp]{ncdf2adp()}} returns an \linkS4class{adp} object.
}
\description{
This works by calling \code{\link[=ncdf2oce]{ncdf2oce()}} and then using \code{\link[=class]{class()}} on
the result to make it be of subclass \code{"adp"}.
}
\examples{
library(ocencdf)

# Example with an adp file from oce package
data(adp, package="oce")
summary(adp)
plot(adp)
# Transfer to netcdf and back to see if results make sense
oce2ncdf(adp, ncfile="adp.nc")
ADP <- ncdf2adp("adp.nc")
summary(ADP)
plot(ADP)

# Remove temporary file
file.remove("adp.nc")

}
\seealso{
Other things related to adp data: 
\code{\link{adp2ncdf}()}
}
\author{
Dan Kelley
}
\concept{things related to adp data}
