% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\name{ncdf2adv}
\alias{ncdf2adv}
\title{Read a NetCDF file and create an adv object}
\usage{
ncdf2adv(ncfile = NULL, varTable = NULL, debug = 0)
}
\arguments{
\item{ncfile}{character value naming the input file.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=ncdf2adv]{ncdf2adv()}} returns an \linkS4class{adv} object.
}
\description{
This works by calling \code{\link[=ncdf2oce]{ncdf2oce()}} and then using \code{\link[=class]{class()}} on
the result to make it be of subclass \code{"adv"}.  This is intended
to work with NetCDF files created with \code{\link[=adv2ncdf]{adv2ncdf()}}, which embeds
sufficient information in the file to permit \code{\link[=ncdf2adv]{ncdf2adv()}} to
reconstruct the original adv object. See the documentation
for \code{\link[=adv2ncdf]{adv2ncdf()}} to learn more about what it stores, and therefore
what \code{\link[=ncdf2adv]{ncdf2adv()}} attempts to read.
}
\examples{
library(ocencdf)

# Example with an adv file from oce package
data(adv, package = "oce")
summary(adv)
plot(adv)
# Transfer to NetCDF and back to see if results make sense
# Use a temporary nc file to let package pass CRAN checks.
ncfile <- tempfile(pattern = "adv", fileext = ".nc")
oce2ncdf(adv, ncfile = ncfile)
ADV <- ncdf2adv(ncfile)
summary(ADV)
plot(ADV)
file.remove(ncfile)

}
\seealso{
Other things related to adv data: 
\code{\link{adv2ncdf}()}

Other things related to adv data: 
\code{\link{adv2ncdf}()}
}
\author{
Dan Kelley
}
\concept{things related to adv data}
