% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf2oce.R
\name{ncdf2ctd}
\alias{ncdf2ctd}
\title{Read a netcdf file and create a ctd object}
\usage{
ncdf2ctd(ncfile = NULL, varTable = NULL, debug = 0)
}
\arguments{
\item{ncfile}{character value naming the input file.}

\item{varTable}{character value indicating the variable-naming
scheme to be used. This is provided to \code{\link[=read.varTable]{read.varTable()}}, the
documentation of which explains the possibilities for pre-defined
schemes named \code{"argo"} and \code{"whp"} and user-supplied files of
similar format.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=ncdf2ctd]{ncdf2ctd()}} returns an \linkS4class{ctd} object.
}
\description{
Read a netcdf file and create a ctd object
}
\seealso{
Other things related to CTD data: 
\code{\link{ctd2ncdf}()}
}
\author{
Dan Kelley
}
\concept{things related to CTD data}
