% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf2oce.R
\name{ncdf2oce}
\alias{ncdf2oce}
\title{Read a netcdf file and create a general oce object}
\usage{
ncdf2oce(ncfile = NULL, varTable = NULL, debug = 0)
}
\arguments{
\item{ncfile}{character value naming the input file.}

\item{varTable}{character value indicating the variable-naming
scheme to be used. This is provided to \code{\link[=read.varTable]{read.varTable()}}, the
documentation of which explains the possibilities for pre-defined
schemes named \code{"argo"} and \code{"whp"} and user-supplied files of
similar format.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=ncdf2oce]{ncdf2oce()}} returns an \linkS4class{oce} object.
}
\description{
Read a netcdf file, interpreting variable names according to
\code{varTable} (if provided).  The object does \emph{not} get a specialized
oce class, because this is not known within netcdf files.  For
ctd data, try \code{\link[=ncdf2ctd]{ncdf2ctd()}} instead of \code{\link[=ncdf2oce]{ncdf2oce()}}, or wrap the
result of calling the latter in \code{\link[oce:as.ctd]{oce::as.ctd()}}.
}
\author{
Dan Kelley
}
