% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartable.R
\name{ncdfNames2oceNames}
\alias{ncdfNames2oceNames}
\title{Translate netcdf names to oce  names}
\usage{
ncdfNames2oceNames(names, varTable = NULL, debug = 0)
}
\arguments{
\item{names}{vector of character values in oce convention (e.g. "TEMP"
for temperature, if varTable is "argo").}

\item{varTable}{character value indicating the variable-naming
scheme to be used. This is provided to \code{\link[=read.varTable]{read.varTable()}}, the
documentation of which explains the possibilities for pre-defined
schemes named \code{"argo"} and \code{"whp"} and user-supplied files of
similar format.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Translate netcdf names to oce  names
}
\author{
Dan Kelley
}
