% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read.varTable}
\alias{read.varTable}
\title{Read a variable-information table}
\usage{
read.varTable(varTable = "argo")
}
\arguments{
\item{varTable}{character value indicating the name of the table.  If this
ends in \code{.yml}, then it is taken as a file to be read.  Alternatively, it
is taken to be the name of a built-in scheme that is defined in a filename
given by appending \code{.yml} on the end of \code{varTable}, to be found in the
\code{inst/ext_data} source directory, as in the example.  At the moment,
the two provided files correspond to \code{varTable} values of \code{"whp"}
and \code{"argo"}.}
}
\value{
\code{\link[=read.varTable]{read.varTable()}} returns a list that specifies some information
to be stored in netcdf files created by e.g. \code{\link[=ctd2ncdf]{ctd2ncdf()}}.
}
\description{
Entries in the file must look like the following.  Note that not all fields
need to be provided.  The first line in each block is the name of
the field as stored in oce objects.  The sub-fields will all default
if not provided, with \code{units} becoming \code{""}, with \code{name}, \code{long_name}
and \code{standard_name} all defaulting to the oce name.  A value
of 99999.0 is used for the missing value if that is not
provided.
\preformatted{
temperature:
   name: "TEMP"
   units: "degree_Celcius"
   long_name: "Sea temperature in-situ ITS-90 scale"
   standard_name: "sea_water_temperature"
   missing_value: 99999.0
}
}
\examples{
library(ocencdf)
str(read.varTable("whp"))

}
\author{
Dan Kelley
}
