% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{adp2ncdf}
\alias{adp2ncdf}
\title{Save an adp object to a netcdf file}
\usage{
adp2ncdf(x, varTable = NULL, ncfile = NULL, debug = 0)
}
\arguments{
\item{x}{an oce object of class \code{adp}, as created by e.g. \code{\link[oce:read.adp]{oce::read.adp()}}.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Given an \code{adp} object created by the \code{oce} package, this function
creates a netcdf file that can later by read by \code{\link[=ncdf2adp]{ncdf2adp()}} to approximately
reproduce the original contents.
}
\details{
Note that \code{\link[=adp2ncdf]{adp2ncdf()}} defaults \code{varTable} to \code{"adp"}.

The entire contents of the metadata slot are saved in the global attribute named
\code{"metadata"}, in a JSON format.  The JSON material is developed with
\code{\link[=metadata2json]{metadata2json()}}, which yields a value that can be decoded with
\code{\link[=json2metadata]{json2metadata()}}.
In addition, the following metadata
items are saved as individual global attributes:
\code{"beamAngle"},
\code{"frequency"},
\code{"instrumentType"},
\code{"instrumentSubtype"},
\code{"numberOfBeams"},
\code{"numberOfBeams"},
and
\code{"oceCoordinate"}.
}
\examples{
library(ocencdf)

# Example with an adp file from oce package
data(adp, package="oce")
summary(adp)
plot(adp)
# Transfer to netcdf and back to see if results make sense
oce2ncdf(adp, ncfile="adp.nc")
ADP <- ncdf2adp("adp.nc")
summary(ADP)
plot(ADP)

# Remove temporary file
file.remove("adp.nc")

}
\seealso{
Other things related to adp data: 
\code{\link{ncdf2adp}()}
}
\author{
Dan Kelley
}
\concept{things related to adp data}
