% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{adp2ncdf}
\alias{adp2ncdf}
\title{Save an ADP object to a NetCDF file}
\usage{
adp2ncdf(x, varTable = NULL, ncfile = NULL, force_v4 = TRUE, debug = 0)
}
\arguments{
\item{x}{an \code{oce} object of class \code{adp}, as created by e.g. \code{\link[oce:read.adp]{oce::read.adp()}}.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{force_v4}{logical value which controls the NetCDF file version during
the \link[ncdf4]{nc_create} step. The default here is TRUE, whereas the
\link[ncdf4]{ncdf4-package} defaults to FALSE (ensuring that the NetCDF
file is compatible with NetCDF v3). Some features, including large data
sizes, may require v4.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Given an \code{adp} object created by the \code{oce} package, this function
creates a NetCDF file that can later by read by \code{\link[=ncdf2adp]{ncdf2adp()}} to approximately
reproduce the original contents.
}
\details{
Note that \code{\link[=adp2ncdf]{adp2ncdf()}} defaults \code{varTable} to \code{"adp"}.

The entire contents of the metadata slot are saved in the global attribute named
\code{"metadata"}, in a JSON format.  The JSON material is developed with
\code{\link[=metadata2json]{metadata2json()}}, which yields a value that can be decoded with
\code{\link[=json2metadata]{json2metadata()}}.
In addition, the following metadata
items are saved as individual global attributes:
\code{"beamAngle"},
\code{"frequency"},
\code{"instrumentType"},
\code{"instrumentSubtype"},
\code{"numberOfBeams"},
\code{"numberOfBeams"},
and
\code{"oceCoordinate"}.
}
\examples{
library(ocencdf)

# Example with an ADP file from `oce` package
data(adp, package = "oce")
summary(adp)
plot(adp)
# Transfer to NetCDF and back to see if results make sense.
# Use a temporary nc file to let package pass CRAN checks.
ncfile <- tempfile(pattern = "adp", fileext = ".nc")
oce2ncdf(adp, ncfile = ncfile)
ADP <- ncdf2adp(ncfile)
summary(ADP)
plot(ADP)
file.remove(ncfile)

}
\seealso{
Other things related to adp data: 
\code{\link{ncdf2adp}()}
}
\author{
Dan Kelley and Clark Richards
}
\concept{things related to adp data}
