% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd2ncdf}
\alias{ctd2ncdf}
\title{Save a ctd object to a NetCDF file}
\usage{
ctd2ncdf(x, varTable = NULL, ncfile = NULL, force_v4 = TRUE, debug = 0)
}
\arguments{
\item{x}{an oce object of class \code{ctd}, as created by e.g. \code{\link[oce:as.ctd]{oce::as.ctd()}}
or \code{\link[oce:read.ctd]{oce::read.ctd()}}.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{force_v4}{logical value which controls the NetCDF file version during
the \link[ncdf4]{nc_create} step. The default here is TRUE, whereas the
\link[ncdf4]{ncdf4-package} defaults to FALSE (ensuring that the NetCDF
file is compatible with NetCDF v3). Some features, including large data
sizes, may require v4.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
This creates a NetCDF file in a convention that permits later reading by
\code{\link[=ncdf2ctd]{ncdf2ctd()}}, and that may be convenient for other purposes as well.
}
\details{
Note that \code{\link[=ctd2ncdf]{ctd2ncdf()}} defaults \code{varTable} to \code{"argo"}.

The contents of the \code{data} slot of the oce object \code{x} are as NetCDF
data items.  If flags are present in the \code{metadata} slot, they are
also saved as data, with names ending in \verb{_QC}.

In addition to storage in the NetCDF data section, several attributes
are saved as well. These include units for the data, which are tied
to the corresponding variables.  The entire \code{metadata} slot is stored
as a global attribute named \code{metadata}, so that a later call to
\code{\link[=ncdf2ctd]{ncdf2ctd()}} will be able to recover the information, using an
\code{eval(parse(text=))} construct.  As an aid to processing in other
languages, the following \code{metadata} items are stored as
individual global attributes: \code{"latitude"}, \code{"longitude"},
\code{"startTime"} and \code{"station"}.
}
\examples{
library(ocencdf)

# example 1: a ctd file without per-variable QC flags
data(ctd, package="oce")
oce2ncdf(ctd, ncfile="ctd.nc")
CTD <- as.ctd(ncdf2oce("ctd.nc"))
summary(CTD)
plot(CTD)

# example 2: a ctd file with per-variable QC flags
data(section, package="oce")
stn <- section[["station", 100]]
oce2ncdf(stn, ncfile="stn.nc")
STN <- as.ctd(ncdf2oce("stn.nc"))
summary(STN)
plot(STN)

# Remove temporary files
file.remove("ctd.nc")
file.remove("stn.nc")

}
\seealso{
Other things related to CTD data: 
\code{\link{ncdf2ctd}()}
}
\author{
Dan Kelley and Clark Richards
}
\concept{things related to CTD data}
