% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf2oce.R
\name{ncdf2oce}
\alias{ncdf2oce}
\title{Read a netcdf file and create a general oce object}
\usage{
ncdf2oce(ncfile = NULL, debug = 0)
}
\arguments{
\item{ncfile}{character value naming the input file.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=ncdf2oce]{ncdf2oce()}} returns an \linkS4class{oce} object.
}
\description{
This reads a netcdf file, interpreting units but not renaming
variables or placing quality-control information in the \code{metadata}
slot.  To get renaming and quality-control handling, call
another oce function as appropriate, e.g. \code{\link[=as.ctd]{as.ctd()}} is
used in the examples provided in the \code{\link[=ctd2ncdf]{ctd2ncdf()}} documentation.
}
\author{
Dan Kelley
}
