% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartable.R
\name{ncdfNames2oceNames}
\alias{ncdfNames2oceNames}
\title{Translate NetCDF names to oce  names}
\usage{
ncdfNames2oceNames(names, varTable = NULL, debug = 0)
}
\arguments{
\item{names}{vector of character values in oce convention (e.g. "TEMP"
for temperature, if varTable is "argo").}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
Translate NetCDF names to oce  names
}
\author{
Dan Kelley
}
