% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce2ncdf.R
\name{oce2ncdf}
\alias{oce2ncdf}
\title{Save an oce-class object as a NetCDF file.}
\usage{
oce2ncdf(x, varTable = NULL, ncfile = NULL, debug = 0)
}
\arguments{
\item{x}{an oce-class object. The subclass of this object determines
which lower-level function \code{\link[=oce2ncdf]{oce2ncdf()}} calls, e.g. if \code{x} is of
\code{ctd} class, then \code{\link[=ctd2ncdf]{ctd2ncdf()}} is called.}

\item{varTable}{character value indicating the variable-naming
scheme to be used, which is passed to \code{\link[=read.varTable]{read.varTable()}} to set
up variable names, units, etc.}

\item{ncfile}{character value naming the output file.  Use NULL
for a file name to be created automatically (e.g. \code{ctd.nc} for
a CTD object).}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\description{
\code{oce2ncdf()} works by determining the class of its first argument,
and then dispatching to an internal function, as appropriate.
}
\author{
Dan Kelley
}
