% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartable.R
\name{read.varTable}
\alias{read.varTable}
\title{Read a variable-information table}
\usage{
read.varTable(varTable = "argo", debug = 0)
}
\arguments{
\item{varTable}{character value indicating the name of the table.  If this
ends in \code{.yml}, then it is taken as a file to be read.  Alternatively, it
is taken to be the name of a built-in scheme that is defined in a filename
given by appending \code{.yml} on the end of \code{varTable}, to be found in the
\code{inst/ext_data} source directory, as in the example.  At the moment,
the two provided files correspond to \code{varTable} values of \code{"whp"}
and \code{"argo"}.}

\item{debug}{integer, 0 (the default) for quiet action apart
from messages and warnings, or any larger value to see more
output that describes the processing steps.}
}
\value{
\code{\link[=read.varTable]{read.varTable()}} returns a list that specifies some information
to be stored in netcdf files created by e.g. \code{\link[=ctd2ncdf]{ctd2ncdf()}}.
}
\description{
This function, meant for internal use by the package, uses
\code{\link[yaml:yaml.load]{yaml::yaml.load_file()}} to read YAML files that describe the
output netcdf format created by e.g. \code{\link[=ctd2ncdf]{ctd2ncdf()}}. Users wishing
to define such files for their own use should follow the pattern
of the source directory \code{inst/extdata/argo.yml}.
}
\examples{
library(ocencdf)
str(read.varTable("whp"))

}
\author{
Dan Kelley
}
